Rename workDefinition reducer to workDefinitionEntry
I think it's confusing with the name workDefinition, when it contains a lot of redux meta data together with the 'real' workDefinition, which is located in the 'entry' attribute.
