+++ src/sentry/static/sentry/app/redux/reducers/index.js
@@ -6,7 +6,7 @@ import processDefinition from './processDefinition';
 import savedSearch from './savedSearch';
 import substanceSearchEntry from './substanceSearchEntry';
 import tag from './tag';
-import workDefinition from './workDefinition';
+import workDefinitionEntry from 'app/redux/reducers/workDefinitionEntry';
 import workBatch from './workBatch';
 import workBatchDetails from './workBatchDetails';
 import projectSearchEntry from './projectSearchEntry';
@@ -23,7 +23,7 @@ export default combineReducers({
   savedSearch,
   substanceSearchEntry,
   tag,
-  workDefinition,
+  workDefinitionEntry,
   workBatch,
   workBatchDetails,
   projectSearchEntry,
