+++ tests/js/spec/redux/reducers/workDefinition.spec.js
@@ -1,10 +1,10 @@
 import {Set} from 'immutable';
-import workDefinition, {initialState} from 'app/redux/reducers/workDefinition';
-import {workDefinitionActions} from 'app/redux/actions/workDefinition';
+import workDefinitionEntry, {initialState} from 'app/redux/reducers/workDefinitionEntry';
+import {workDefinitionActions} from 'app/redux/actions/workDefinitionEntry';
 
 describe('workDefinition reducer, entry protocol', () => {
   it('has expected state after requesting an item', () => {
-    const requested = workDefinition(
+    const requested = workDefinitionEntry(
       initialState,
       workDefinitionActions.getRequest('id')
     );
@@ -17,12 +17,15 @@ describe('workDefinition reducer, entry protocol', () => {
   });
 
   it('expected state after get success', () => {
-    const requested = workDefinition(
+    const requested = workDefinitionEntry(
       initialState,
       workDefinitionActions.getRequest('id')
     );
     const entry = TestStubs.WorkDefinition('work-definition-name');
-    const succeeded = workDefinition(requested, workDefinitionActions.getSuccess(entry));
+    const succeeded = workDefinitionEntry(
+      requested,
+      workDefinitionActions.getSuccess(entry)
+    );
 
     const expected = {
       loadingDetails: false,
